requirements.txt
aiogram==3.10.0
boto3==1.34.0
botocore==1.34.0
gspread==5.12.0
progress==1.5
python-dotenv==1.0.1